third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGCircleElement-svgdom-cx-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'cx' property of the SVGCircleElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS circleElement.cx.baseVal.value is -50
PASS circleElement.cx.baseVal.value is 150
PASS successfullyParsed is true

TEST COMPLETE
